Moving HttpClientBuilder to protocol.http package
