+++ src/main/java/io/reactivex/netty/RxNetty.java
@@ -20,11 +20,11 @@ import io.netty.handler.codec.http.FullHttpResponse;
 import io.netty.handler.codec.http.HttpObject;
 import io.netty.handler.codec.http.HttpRequest;
 import io.reactivex.netty.client.ClientBuilder;
-import io.reactivex.netty.client.HttpClientBuilder;
 import io.reactivex.netty.client.RxClient;
 import io.reactivex.netty.pipeline.PipelineConfigurator;
 import io.reactivex.netty.pipeline.PipelineConfigurators;
 import io.reactivex.netty.protocol.http.HttpClient;
+import io.reactivex.netty.protocol.http.HttpClientBuilder;
 import io.reactivex.netty.protocol.http.HttpServer;
 import io.reactivex.netty.protocol.http.HttpServerBuilder;
 import io.reactivex.netty.protocol.text.sse.SSEEvent;
